Доработал api namespace_service

+++ namespace_service/namespace_service/routers.py
@@ -1,17 +1,25 @@
 from typing import List
 from uuid import UUID
 
-from fastapi import APIRouter, status, Depends
+from fastapi import APIRouter, Depends, HTTPException, status
 
-from namespace_service.db.repository import NamespaceRepository, get_namespace_repo
 from namespace_service.auth import get_current_user, get_superuser
-from namespace_service.schemas import NamespaceModel, InputNamespace, \
-    UserNamespacesPage, NamespaceModelWithUsers, NamespaceUsers, UserNamespaces, UserModel, UserNamespace
+from namespace_service.db.repository import NamespaceRepository, get_namespace_repo
+from namespace_service.schemas import (
+    InputNamespace,
+    NamespaceModel,
+    NamespaceModelWithUsers,
+    NamespaceUsers,
+    UserModel,
+    UserNamespace,
+    UserNamespaces,
+    UserNamespacesPage,
+)
 
 router = APIRouter()
 
 
-@router.get('/namespaces', status_code=status.HTTP_200_OK, response_model=UserNamespacesPage)
+@router.get('', status_code=status.HTTP_200_OK, response_model=UserNamespacesPage)
 async def get_user_namespaces(
     page: int = 1,
     size: int = 100,
@@ -20,7 +28,9 @@ async def get_user_namespaces(
 ) -> UserNamespacesPage:
     user_namespaces: UserNamespaces = await repo.get_user_namespaces(current_user.id)
 
-    namespaces: List[NamespaceModel] = [await repo.get_namespace(namespace_id) for namespace_id in user_namespaces.namespaces]
+    namespaces: List[NamespaceModel] = [
+        await repo.get_namespace(namespace_id) for namespace_id in user_namespaces.namespaces
+    ]
 
     result_count = len(user_namespaces.namespaces)
 
@@ -35,40 +45,46 @@ async def get_user_namespaces(
     return UserNamespacesPage(page=page, size=size, total_elements=result_count, items=namespaces)
 
 
-@router.get('/namespaces/{namespace_id}', status_code=status.HTTP_200_OK, response_model=NamespaceModelWithUsers)
+@router.get('/{namespace_id}', status_code=status.HTTP_200_OK, response_model=NamespaceModelWithUsers)
 async def get_namespace(
     namespace_id: UUID,
     repo: NamespaceRepository = Depends(get_namespace_repo),
     current_user: UserModel = Depends(get_current_user),
 ) -> NamespaceModelWithUsers:
     namespace_model: NamespaceModel = await repo.get_namespace(namespace_id)
+    if namespace_model.owner_id != current_user.id and not current_user.is_superuser:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='Only namespace owner can get info about it')
     namespace_users: NamespaceUsers = await repo.get_namespace_users(namespace_model.id)
     return NamespaceModelWithUsers(**namespace_model.dict(), users=namespace_users.users)
 
 
-@router.post('/namespaces', status_code=status.HTTP_201_CREATED, response_model=NamespaceModel)
+@router.post('', status_code=status.HTTP_201_CREATED, response_model=NamespaceModel)
 async def create_namespace(
     namespace: InputNamespace,
     repo: NamespaceRepository = Depends(get_namespace_repo),
-    current_user: UserModel = Depends(get_superuser),
+    current_user: UserModel = Depends(get_current_user),
 ) -> NamespaceModel:
-    return await repo.create_namespace(namespace)
+    namespace = await repo.create_namespace(namespace, current_user.id)
+    await repo.add_user_to_namespace(namespace.id, current_user.id)
+    return namespace
 
 
-@router.delete('/namespaces/{namespace_id}', status_code=status.HTTP_204_NO_CONTENT, response_model=None)
+@router.delete('/{namespace_id}', status_code=status.HTTP_204_NO_CONTENT, response_model=None)
 async def delete_namespace(
     namespace_id: UUID,
     repo: NamespaceRepository = Depends(get_namespace_repo),
-    current_user: UserModel = Depends(get_superuser),
+    current_user: UserModel = Depends(get_current_user),
 ) -> None:
+    namespace: NamespaceModel = await repo.get_namespace(namespace_id)
+    if namespace.owner_id != current_user.id and not current_user.is_superuser:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='Only namespace owner can delete it')
+    namespace_users: NamespaceUsers = await repo.get_namespace_users(namespace_id)
+    for user_id in namespace_users.users:
+        await repo.delete_user_from_namespace(namespace_id, user_id)
     await repo.delete_namespace(namespace_id)
 
 
-@router.get(
-    'namespaces/{namespace_id}/users/{user_id}',
-    status_code=status.HTTP_200_OK,
-    response_model=UserNamespace
-)
+@router.get('/{namespace_id}/users/{user_id}', status_code=status.HTTP_200_OK, response_model=UserNamespace)
 async def get_user_in_namespace(
     user_id: UUID,
     namespace_id: UUID,
@@ -78,11 +94,7 @@ async def get_user_in_namespace(
     return await repo.get_namespace_user(user_id, namespace_id)
 
 
-@router.get(
-    'namespaces/{namespace_id}/users/me',
-    status_code=status.HTTP_200_OK,
-    response_model=UserNamespace
-)
+@router.get('/{namespace_id}/users/me', status_code=status.HTTP_200_OK, response_model=UserNamespace)
 async def get_self_user_in_namespace(
     namespace_id: UUID,
     repo: NamespaceRepository = Depends(get_namespace_repo),
@@ -91,29 +103,35 @@ async def get_self_user_in_namespace(
     return await repo.get_namespace_user(current_user.id, namespace_id)
 
 
-@router.post(
-    'namespaces/{namespace_id}/users/{user_id}',
-    status_code=status.HTTP_201_CREATED,
-    response_model=UserNamespacesPage
-)
+@router.post('/{namespace_id}/users/{user_id}', status_code=status.HTTP_204_NO_CONTENT)
 async def add_user_to_namespace(
     user_id: UUID,
     namespace_id: UUID,
     repo: NamespaceRepository = Depends(get_namespace_repo),
     current_user: UserModel = Depends(get_superuser),
 ) -> None:
+    namespace: NamespaceModel = await repo.get_namespace(namespace_id)
+    if namespace.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail='Only namespace owner can add new namespace memberships'
+        )
     await repo.add_user_to_namespace(namespace_id, user_id)
 
 
-@router.delete(
-    '/namespaces/{namespace_id}/users/{user_id}',
-    status_code=status.HTTP_204_NO_CONTENT,
-    response_model=None
-)
+@router.delete('/{namespace_id}/users/{user_id}', status_code=status.HTTP_204_NO_CONTENT)
 async def delete_user_from_namespace(
     user_id: UUID,
     namespace_id: UUID,
     repo: NamespaceRepository = Depends(get_namespace_repo),
     current_user: UserModel = Depends(get_superuser),
 ) -> None:
+    namespace: NamespaceModel = await repo.get_namespace(namespace_id)
+    if namespace.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail='Only namespace owner can delete namespace memberships'
+        )
+    if namespace.owner_id == user_id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail='Can not delete namespace owner from namespace'
+        )
     await repo.delete_user_from_namespace(namespace_id, user_id)
